SCHREGISTER-66 Доработка планировщиков I_SCHR_1, I_SCHR_3, I_SCHR_5, I_SCHR_7, I_SCHR_9, I_SCHR_11

+++ application/schr-scheduler/erp-change-patient-policies-import-service/erp-change-patient-domain/src/main/java/moscow/ptnl/app/ecppis/deserilizer/ErpChangePatientPoliciesDeserializer.java
@@ -64,9 +64,9 @@ public class ErpChangePatientPoliciesDeserializer implements Function<String, Er
         TypeRef<List<EntityData>> typeRef = new TypeRef<List<EntityData>>() {};
 
         return new ErpChangePatientPolicies(
-                extractSingle(value, "$.id", Integer.class),
+                extractSingle(value, "$.id", Long.class),
                 extractSingle(value, "$.operationDate", LocalDateTime.class),
-                extractSingle(value, "$.emiasId", Integer.class),
+                extractSingle(value, "$.emiasId", Long.class),
                 JsonPath.read(value, "$.uklErp"),
                 JsonPath.read(value, "$.patientType"),
                 JsonPath.read(value, "$.patientRecStatus"),
